+++ runner_tests.sh
@@ -2,15 +2,17 @@ export LOGIN=$1
 export PASSWORD=$2
 echo $LOGIN $PASSWORD
 
-pytest --numprocesses=1 testsbase/test_person.py
-pytest --numprocesses=1 testsbase/test_collection.py
+pytest testsbase/test_collection.py
+pytest testsbase/test_person.py
 
-pytest --numprocesses=1 testsbase/test_login.py
-pytest --numprocesses=1 testsbase/test_signup.py
-pytest --numprocesses=1 testsbase/test_profile.py
+pytest testsbase/test_search.py
+pytest testsbase/test_film_unauth.py
+pytest testsbase/test.py
 
-pytest --numprocesses=1 testsbase/test_navigation.py
-pytest --numprocesses=1 testsbase/test_mainpage.py
-pytest --numprocesses=1 testsbase/test_premierepage.py
+pytest testsbase/test_login.py
+pytest testsbase/test_signup.py
+pytest testsbase/test_profile.py
 
-pytest --numprocesses=1 testsbase/test.py
+pytest testsbase/test_navigation.py
+pytest testsbase/test_mainpage.py
+pytest testsbase/test_premierepage.py
