+++ testsbase/test_film_unauth.py
@@ -1,0 +1,70 @@
+import time
+
+from selenium.webdriver.common.keys import Keys
+from selenium.webdriver.common.by import By
+
+from driver import dvr
+from base_page import BasePage
+from const import *
+from helper_auth import needed_auth
+
+class TestFilmPageUnauthorized(BasePage):
+    X_BUTTON_PLUS = '/html/body/div/div/div[1]/div/div[2]/div[7]/div/a/div'
+    X_BUTTON_BOOKMARK = '/html/body/div/div/div[1]/div/div[2]/div[7]/a[2]/div'
+    X_BUTTON_TRAILER = '/html/body/div/div/div[1]/div/div[2]/div[7]/a[1]/div'
+    X_PRODUCER = '/html/body/div/div/div[1]/div/div[2]/div[5]/div[2]'
+    X_PRODUCER_ON_PAGE = '/html/body/div/div/div[1]/div/div/div/div[1]'
+    X_TOSTER = '/html/body/div/div[2]'
+    X_TRAILER_CONTAINER = '/html/body/div/div[1]/div/div/iframe'
+    X_BUTTON_RATE_9 = '/html/body/div/div/div[2]/div[2]/div[2]/form/div/div[2]/div[1]/button[2]'
+    X_BUTTON_RATE_1 = '/html/body/div/div/div[2]/div[2]/div[2]/form/div/div[2]/div[1]/button[10]'
+    X_BUTTON_RATE_CONTAINER = '/html/body/div/div/div[2]/div[2]/div[2]/form/div/div[2]/div[1]'
+    X_BUTTON_RATE_DELETE = '/html/body/div/div/div[2]/div[2]/div[2]/form/div/div[4]/button'
+    X_BUTTON_REVIEW = '/html/body/div/div/div[4]/div[1]/div/div/div[1]/a/div'
+    X_BUTTON_REVIEW_CONTAINER = '/html/body/div/div[1]/div/div/div/div'
+    X_BUTTON_CHOOSE_TYPE = '/html/body/div/div[1]/div/div/div/div/form/div[1]/div/div'
+    X_BUTTON_NEGATIVE_TYPE = '/html/body/div/div[1]/div/div/div/div/form/div[1]/div/ul/li[3]'
+    X_BUTTON_NEUTRAL_TYPE = '/html/body/div/div[1]/div/div/div/div/form/div[1]/div/ul/li[2]'
+    X_BUTTON_POSITIVE_TYPE = '/html/body/div/div[1]/div/div/div/div/form/div[1]/div/ul/li[1]'
+    X_INPUT_TITLE = '/html/body/div/div[1]/div/div/div/div/form/div[2]/input'
+    X_INPUT_TEXT = '/html/body/div/div[1]/div/div/div/div/form/div[3]/textarea'
+    X_BUTTON_SEND_REVIEW = '/html/body/div/div[1]/div/div/div/div/form/button'
+    X_UPPER_REVIEW = '/html/body/div/div/div[4]/div[1]/div/div/div[2]/div[1]/div'
+    X_UPPER_REVIEW_TITLE = '/html/body/div/div/div[4]/div[1]/div/div/div[2]/div[1]/div/div[2]'
+    X_UPPER_REVIEW_TEXT = '/html/body/div/div/div[4]/div[1]/div/div/div[2]/div[1]/div/div[3]'
+    TEXT = '200IQ text'
+    TITLE = 'title'
+
+    def test_click_producer(self):
+        self.render(f'{DOMAIN}/film/1')
+
+        prod = self.find((By.XPATH, self.X_PRODUCER), 3).text.strip()
+        self.find((By.XPATH, self.X_PRODUCER), 5).click()
+
+        prod_on_page = self.find((By.XPATH, self.X_PRODUCER_ON_PAGE), 3).text.strip()
+        if prod != prod_on_page:
+            raise Exception("names does not equal", prod, prod_on_page)
+
+    def test_click_plus_unauth(self):
+        self.render(f'{DOMAIN}/film/1')
+
+        self.find((By.XPATH, self.X_BUTTON_PLUS)).click()
+        self.find((By.XPATH, self.X_TOSTER))
+        self.wait_hide((By.XPATH, self.X_TOSTER))
+
+        self.find((By.XPATH, self.X_BUTTON_BOOKMARK)).click()
+        self.find((By.XPATH, self.X_TOSTER))
+        self.wait_hide((By.XPATH, self.X_TOSTER))
+
+    def test_click_trailer(self):
+        self.render(f'{DOMAIN}/film/1')
+        self.find((By.XPATH, self.X_BUTTON_TRAILER)).click()
+        self.find((By.XPATH, self.X_TRAILER_CONTAINER))
+
+    def test_click_review_unauth(self):
+        self.render(f'{DOMAIN}/film/1')
+
+        self.find((By.XPATH, self.X_BUTTON_REVIEW)).click()
+
+        self.find((By.XPATH, self.X_TOSTER))
+        self.wait_hide((By.XPATH, self.X_TOSTER))
