+++ testsbase/test_search.py
@@ -1,0 +1,63 @@
+import time
+
+from selenium.webdriver.common.keys import Keys
+from selenium.webdriver.common.by import By
+
+from driver import dvr
+from base_page import BasePage
+from const import *
+from helper_auth import needed_auth
+
+class TestSearch(BasePage):
+    X_INPUT_SEARCH = '/html/body/div/header/form/input'
+    X_BUTTON_RETURN_MAIN = '/html/body/div/div/div[2]/div[4]/div'
+    X_TITLE_SEARCH = '/html/body/div/div/div[1]'
+    TITLE_SEARCH = 'Результаты поиска'
+    ALL_GROUP_SEARCH = 'a'
+    X_SEARCH_GROUP_FILMS = '/html/body/div/div/div[2]/div[1]/div/div/div[1]'
+    SEARCH_GROUP_FILMS = 'Найденные фильмы:'
+    X_SEARCH_GROUP_SERAILS = '/html/body/div/div/div[2]/div[2]/div/div/div[1]'
+    SEARCH_GROUP_SERAILS = 'Найденные сериалы:'
+    X_SEARCH_GROUP_PERSONS = '/html/body/div/div/div[2]/div[3]/div/div/div[1]'
+    SEARCH_GROUP_PERSONS = 'Найденные имена:'
+    X_FIRST_FOUNDED_FILM = '/html/body/div/div/div[2]/div[1]/div/div/div[2]/div[1]/div/div[1]/div[1]/div[1]'
+    X_FIRST_FOUNDED_TITLE = '/html/body/div/div/div[2]/div[1]/div/div/div[2]/div[1]/div/div[1]/div[2]/div[1]/div[1]'
+    X_TITLE_ON_FILM_PAGE = '/html/body/div/div/div[1]/div/div[2]/div[1]'
+
+    def test_check_exists(self):
+        self.render(DOMAIN)
+        self.find((By.XPATH, self.X_INPUT_SEARCH)).send_keys(Keys.ENTER)
+        title = self.find((By.XPATH, self.X_TITLE_SEARCH)).text
+        if title != self.TITLE_SEARCH:
+            raise Exception("stings does not equal", title, self.TITLE_SEARCH)
+
+    def test_check_empty(self):
+        self.render(DOMAIN)
+        self.find((By.XPATH, self.X_INPUT_SEARCH)).send_keys(Keys.ENTER)
+        self.find((By.XPATH, self.X_BUTTON_RETURN_MAIN)).click()
+
+    def test_check_group_category(self):
+        self.render(DOMAIN)
+        self.find((By.XPATH, self.X_INPUT_SEARCH)).send_keys(self.ALL_GROUP_SEARCH, Keys.ENTER)
+
+        films_field = self.find((By.XPATH, self.X_SEARCH_GROUP_FILMS)).text
+        if films_field != self.SEARCH_GROUP_FILMS:
+            raise Exception("stings does not equal", films_field, self.SEARCH_GROUP_FILMS)
+
+        serials_field = self.find((By.XPATH, self.X_SEARCH_GROUP_SERAILS)).text
+        if serials_field != self.SEARCH_GROUP_SERAILS:
+            raise Exception("stings does not equal", serials_field, self.SEARCH_GROUP_SERAILS)
+
+        persons_field = self.find((By.XPATH, self.X_SEARCH_GROUP_PERSONS)).text
+        if persons_field != self.SEARCH_GROUP_PERSONS:
+            raise Exception("stings does not equal", persons_field, self.SEARCH_GROUP_PERSONS)
+
+    def test_correct_results(self):
+        self.render(DOMAIN)
+        self.find((By.XPATH, self.X_INPUT_SEARCH)).send_keys(self.ALL_GROUP_SEARCH, Keys.ENTER)
+
+        films_field = self.find((By.XPATH, self.X_FIRST_FOUNDED_TITLE)).text
+        self.find((By.XPATH, self.X_FIRST_FOUNDED_FILM)).click()
+        film_title = self.find((By.XPATH, self.X_TITLE_ON_FILM_PAGE)).text
+        if films_field != film_title:
+            raise Exception("stings does not equal", films_field, film_title)
